fix: use u128 intermediate math to prevent overflow in share calculations
Deposit and withdraw share calculations overflowed u64 when
shares * deposits exceeded ~1.8e19. Cast to u128 for the intermediate
multiply before dividing back to u64.

Also adds comprehensive E2E test (6 steps: init, deposit, atomic
borrow+repay, fee verification, LP withdraw, security checks) and
devnet program config.

+++ programs/solana-flash-loan/src/instructions/deposit.rs
@@ -53,11 +53,12 @@ pub fn handle_deposit(ctx: Context<Deposit>, amount: u64) -> Result<()> {
     let shares = if pool.total_shares == 0 {
         amount
     } else {
-        amount
-            .checked_mul(pool.total_shares)
+        // Use u128 intermediate to avoid overflow: (amount * total_shares) / total_deposits
+        (amount as u128)
+            .checked_mul(pool.total_shares as u128)
             .ok_or(FlashLoanError::MathOverflow)?
-            .checked_div(pool.total_deposits)
-            .ok_or(FlashLoanError::MathOverflow)?
+            .checked_div(pool.total_deposits as u128)
+            .ok_or(FlashLoanError::MathOverflow)? as u64
     };
 
     // Transfer tokens from depositor to vault
